Handle include LoadError
